refactor: centralize createReqRes utility
